fixed error with box size changed int to float

+++ singleparticle/convert.py
@@ -94,14 +94,14 @@ def write_csv(filename, data):
         csvwriter.writerows(data)
 
 
-def read_coordinate3D(csv_file: str) -> Iterator[Tuple[Coordinate3D, int]]:
+def read_coordinate3D(csv_file: str) -> Iterator[Tuple[Coordinate3D, float]]:
     with open(csv_file, "r") as f:
         data = csv.reader(f)
         next(data)
         for row in data:
             coord = Coordinate3D()
             coord.setPosition(float(row[1]), float(row[2]), float(row[3]))
-            yield coord, int(float(row[4]))
+            yield coord, float(row[4])
 
 
 def save_translations(coords: SetOfCoordinates3D, csv_file: str):
